Fixed a bug in the document normalization process
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/trunk@14006 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ weka/src/main/java/weka/classifiers/bayes/NaiveBayesMultinomialText.java
@@ -310,7 +310,7 @@ public class NaiveBayesMultinomialText extends AbstractClassifier implements
 
       LinkedHashMap<String, Count> dictForClass = m_probOfWordGivenClass.get(i);
 
-      int allWords = 0;
+      double allWords = 0;
       // for document normalization (if in use)
       double iNorm = 0;
       double fv = 0;
